Working on qol features, using cims masterplan, OK up to Scene 4

+++ structure/Converter.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.IO;
+using System.Windows.Forms;
 
 namespace ImageUtil.structure
 {
@@ -15,20 +16,52 @@ namespace ImageUtil.structure
 
         // Parent convert, uses default System.Drawing Image library
         // Usable formats in ./Program.cs
-        public virtual void convert(List<String> files, bool keepFiles)
+        public virtual List<bool> convert(List<String> files, bool keepFiles)
         {
-           
+            List<bool> result = new List<bool>();
+            List<String> files2 = files;
+            files2.RemoveAll(x => x.EndsWith(this.toFormat));
+
+            int successful = 0;
+
+            if (!keepFiles)
+            {
+                DialogResult overwrite = MessageBox.Show("This action will overwrite existing files. Continue?", "Overwrite", MessageBoxButtons.YesNo);
+                if ( overwrite == DialogResult.No) { return new List<bool>(); }
+            }
+
             foreach (String file in files)
             {
-                Console.WriteLine($"The active format during conversion is: {toFormat}");
+                Console.WriteLine($"Files to convert: {files2.Count}");
                 String name = Path.GetFileNameWithoutExtension(file);
-                using (Image sourceImage = Image.FromFile(file))
+                try
                 {
-                    ImageFormat format = (ImageFormat)new ImageFormatConverter().ConvertFromString(toFormat);
-                    sourceImage.Save($"{Path.GetDirectoryName(file)}\\{name}.{toFormat}", format);
+                    using (Image sourceImage = Image.FromFile(file))
+                    {
+                        ImageFormat format = (ImageFormat)new ImageFormatConverter().ConvertFromString(this.toFormat);
+                        sourceImage.Save($"{Path.GetDirectoryName(file)}\\{name}.{this.toFormat}", format);
+                        result.Add(true);
+                        successful++; 
+                    }
+                    if (!keepFiles) { File.Delete(file); }
                 }
-                if (!keepFiles) { File.Delete(file); }
+                catch (Exception ex)
+                {
+                    // Switch with Exception case is not supported in my current version of C#, bummer
+                    // cancel the image convert and display the error message to the user
+                    if ( ex is BadImageFormatException ) { }
+                    if ( ex is NotSupportedException ) { }
+                    if ( ex is System.Security.SecurityException ) { }
+
+                    result.Add(false);
+                }
             }
+            if ( result.All(x => x) )
+            {
+                MessageBox.Show($"Successfully converted {files2.Count} file(s)", "Success", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            }
+            else { MessageBox.Show($"Successfully converted {successful} file(s), failed to convert {files2.Count - successful} file(s)", "Result", MessageBoxButtons.OK, MessageBoxIcon.Information); }
+            return result;
         }
     }
 }
